+++ src/Octoshift/GithubApi.cs
@@ -5,7 +5,7 @@ using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
-using OctoshiftCLI.Data;
+using OctoshiftCLI.Models;
 
 namespace OctoshiftCLI
 {
